Moves transforms to a package and fleshes out the html->cnxml transforms.
There are a lot of name changes here. Some of the logic has changed,
but the abstract behavior is very similar. These changes are likely not
compatible with cnx-upgrade.
